update weight_kg data_type to DECIMAL

+++ schema.sql
@@ -6,7 +6,7 @@ CREATE TABLE animals (
 	date_of_birth DATE,
 	escape_attempts INTEGER,
 	neutered BOOLEAN,
-	weight_kg FLOAT(2)
+	weight_kg DECIMAL
 )
